isolated computeStats into update_stats, call it after computeVaccination
